+++ src/CleanRegex/Composite/ChainedReplace.php
@@ -2,6 +2,8 @@
 namespace TRegx\CleanRegex\Composite;
 
 use TRegx\CleanRegex\Internal\Definition;
+use TRegx\CleanRegex\Internal\GroupKey\WholeMatch;
+use TRegx\CleanRegex\Internal\Replace\AllowAllGroupAware;
 use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\SubjectRs;
 use TRegx\CleanRegex\Internal\Replace\Counting\IgnoreCounting;
 use TRegx\CleanRegex\Internal\Replace\ReplaceReferences;
@@ -56,7 +58,8 @@ class ChainedReplace
 
     private function replaceNext(Definition $definition, string $subject, callable $callback): string
     {
-        $invoker = new ReplacePatternCallbackInvoker($definition, new Subject($subject), -1, $this->substitute, new IgnoreCounting());
+        $invoker = new ReplacePatternCallbackInvoker($definition, new Subject($subject), -1, $this->substitute, new IgnoreCounting(),
+            new AllowAllGroupAware(), new WholeMatch());
         return $invoker->invoke($callback, new MatchStrategy());
     }
 }