+++ src/CleanRegex/Replace/By/ByReplacePattern.php
@@ -1,8 +1,10 @@
 <?php
 namespace TRegx\CleanRegex\Replace\By;
 
+use TRegx\CleanRegex\Internal\Definition;
 use TRegx\CleanRegex\Internal\GroupKey\GroupKey;
 use TRegx\CleanRegex\Internal\GroupKey\WholeMatch;
+use TRegx\CleanRegex\Internal\Model\GroupAware;
 use TRegx\CleanRegex\Internal\Replace\By\GroupFallbackReplacer;
 use TRegx\CleanRegex\Internal\Replace\By\GroupMapper\DictionaryMapper;
 use TRegx\CleanRegex\Internal\Replace\By\GroupMapper\SubstituteFallbackMapper;
@@ -10,6 +12,7 @@ use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\DefaultStrategy;
 use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\LazySubjectRs;
 use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\ThrowMatchRs;
 use TRegx\CleanRegex\Internal\Replace\By\PerformanceEmptyGroupReplace;
+use TRegx\CleanRegex\Internal\Replace\Counting\CountingStrategy;
 use TRegx\CleanRegex\Internal\Replace\Wrapper;
 use TRegx\CleanRegex\Internal\Replace\WrappingMapper;
 use TRegx\CleanRegex\Internal\Subject;
@@ -25,33 +28,56 @@ class ByReplacePattern
     private $subject;
     /** @var PerformanceEmptyGroupReplace */
     private $performanceReplace;
-    /** @var ReplacePatternCallbackInvoker */
-    private $replaceCallbackInvoker;
+    /** @var Definition */
+    private $definition;
+    /** @var int */
+    private $limit;
+    /** @var CountingStrategy */
+    private $countingStrategy;
+    /** @var GroupAware */
+    private $groupAware;
     /** @var Wrapper */
     private $wrapper;
 
-    public function __construct(GroupFallbackReplacer         $fallbackReplacer,
-                                LazySubjectRs                 $substitute,
-                                PerformanceEmptyGroupReplace  $performanceReplace,
-                                ReplacePatternCallbackInvoker $replaceCallbackInvoker,
-                                Subject                       $subject,
-                                Wrapper                       $middlewareMapper)
+    public function __construct(GroupFallbackReplacer        $fallbackReplacer,
+                                LazySubjectRs                $substitute,
+                                PerformanceEmptyGroupReplace $performanceReplace,
+                                Definition                   $definition,
+                                int                          $limit,
+                                CountingStrategy             $countingStrategy,
+                                GroupAware                   $groupAware,
+                                Subject                      $subject,
+                                Wrapper                      $middlewareMapper)
     {
         $this->fallbackReplacer = $fallbackReplacer;
         $this->substitute = $substitute;
         $this->subject = $subject;
         $this->performanceReplace = $performanceReplace;
-        $this->replaceCallbackInvoker = $replaceCallbackInvoker;
+        $this->definition = $definition;
+        $this->limit = $limit;
+        $this->countingStrategy = $countingStrategy;
+        $this->groupAware = $groupAware;
         $this->wrapper = $middlewareMapper;
     }
 
     public function group($nameOrIndex): ByGroupReplacePattern
     {
+        return $this->replaceGroup(GroupKey::of($nameOrIndex));
+    }
+
+    private function replaceGroup(GroupKey $group): ByGroupReplacePattern
+    {
         return new ByGroupReplacePattern(
             $this->fallbackReplacer,
             $this->performanceReplace,
-            $this->replaceCallbackInvoker,
-            GroupKey::of($nameOrIndex),
+            new ReplacePatternCallbackInvoker($this->definition,
+                $this->subject,
+                $this->limit,
+                $this->substitute,
+                $this->countingStrategy,
+                $this->groupAware,
+                $group),
+            $group,
             $this->subject,
             $this->wrapper);
     }