+++ src/CleanRegex/Replace/Callback/ReplacePatternCallbackInvoker.php
@@ -1,10 +1,12 @@
 <?php
 namespace TRegx\CleanRegex\Replace\Callback;
 
+use TRegx\CleanRegex\Exception\NonexistentGroupException;
 use TRegx\CleanRegex\Internal\Definition;
+use TRegx\CleanRegex\Internal\GroupKey\GroupKey;
 use TRegx\CleanRegex\Internal\Match\Base\ApiBase;
 use TRegx\CleanRegex\Internal\Match\MatchAll\LazyMatchAllFactory;
-use TRegx\CleanRegex\Internal\Match\MatchAll\MatchAllFactory;
+use TRegx\CleanRegex\Internal\Model\GroupAware;
 use TRegx\CleanRegex\Internal\Model\LightweightGroupAware;
 use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\SubjectRs;
 use TRegx\CleanRegex\Internal\Replace\Counting\CountingStrategy;
@@ -23,10 +25,19 @@ class ReplacePatternCallbackInvoker
     private $substitute;
     /** @var CountingStrategy */
     private $countingStrategy;
-    /** @var MatchAllFactory */
+    /** @var GroupAware */
+    private $groupAware;
+    /** @var GroupKey */
+    private $group;
     private $allFactory;
 
-    public function __construct(Definition $definition, Subject $subject, int $limit, SubjectRs $substitute, CountingStrategy $countingStrategy)
+    public function __construct(Definition       $definition,
+                                Subject          $subject,
+                                int              $limit,
+                                SubjectRs        $substitute,
+                                CountingStrategy $countingStrategy,
+                                GroupAware       $groupAware,
+                                GroupKey         $group)
     {
         $this->definition = $definition;
         $this->subject = $subject;
@@ -34,6 +45,8 @@ class ReplacePatternCallbackInvoker
         $this->substitute = $substitute;
         $this->countingStrategy = $countingStrategy;
         $this->allFactory = new LazyMatchAllFactory(new ApiBase($definition, $subject));
+        $this->groupAware = $groupAware;
+        $this->group = $group;
     }
 
     public function invoke(callable $callback, ReplaceCallbackArgumentStrategy $strategy): string
@@ -41,9 +54,20 @@ class ReplacePatternCallbackInvoker
         $result = $this->pregReplaceCallback($callback, $replaced, $strategy);
         $this->countingStrategy->count($replaced, new LightweightGroupAware($this->definition));
         if ($replaced === 0) {
-            return $this->substitute->substitute($this->subject) ?? $result;
+            if ($this->groupExists()) {
+                return $this->substitute->substitute($this->subject) ?? $result;
+            }
+            throw new NonexistentGroupException($this->group);
         }
         return $result;
+    }
+
+    private function groupExists(): bool
+    {
+        if ($this->group->nameOrIndex() === 0) {
+            return true;
+        }
+        return $this->groupAware->hasGroup($this->group);
     }
 
     private function pregReplaceCallback(callable $callback, ?int &$replaced, ReplaceCallbackArgumentStrategy $strategy): string
@@ -66,7 +90,7 @@ class ReplacePatternCallbackInvoker
 
     private function createObjectCallback(callable $callback, ReplaceCallbackArgumentStrategy $strategy): callable
     {
-        $object = new ReplaceCallbackObject($callback, $this->subject, $this->allFactory, $this->limit, $strategy);
+        $object = new ReplaceCallbackObject($callback, $this->subject, $this->allFactory, $this->limit, $strategy, $this->groupAware, $this->group);
         return $object->getCallback();
     }
 }