+++ src/CleanRegex/Replace/FocusReplacePattern.php
@@ -5,6 +5,7 @@ use TRegx\CleanRegex\Exception\FocusGroupNotMatchedException;
 use TRegx\CleanRegex\Internal\Definition;
 use TRegx\CleanRegex\Internal\GroupKey\GroupKey;
 use TRegx\CleanRegex\Internal\Match\Base\ApiBase;
+use TRegx\CleanRegex\Internal\Model\LightweightGroupAware;
 use TRegx\CleanRegex\Internal\Replace\By\GroupFallbackReplacer;
 use TRegx\CleanRegex\Internal\Replace\By\GroupMapper\FocusWrapper;
 use TRegx\CleanRegex\Internal\Replace\By\NonReplaced\DefaultStrategy;
@@ -15,7 +16,6 @@ use TRegx\CleanRegex\Internal\Replace\ReferencesReplacer;
 use TRegx\CleanRegex\Internal\Subject;
 use TRegx\CleanRegex\Match\Details\Detail;
 use TRegx\CleanRegex\Replace\By\ByReplacePattern;
-use TRegx\CleanRegex\Replace\Callback\ReplacePatternCallbackInvoker;
 
 class FocusReplacePattern implements SpecificReplacePattern
 {
@@ -92,7 +92,8 @@ class FocusReplacePattern implements SpecificReplacePattern
                 new ApiBase($this->definition, $this->subject)),
             new LazyMessageThrowStrategy(),
             new PerformanceEmptyGroupReplace($this->definition, $this->subject, $this->limit),
-            new ReplacePatternCallbackInvoker($this->definition, $this->subject, $this->limit, new DefaultStrategy(), $this->countingStrategy),
+            $this->definition, $this->limit, $this->countingStrategy,
+            new LightweightGroupAware($this->definition),
             $this->subject,
             new FocusWrapper($this->group));
     }